Changed float to double

+++ jcnet/Functions.java
@@ -3,7 +3,7 @@ import jcnet.backwardfunctions.ScalarBackward;
 public class Functions {
 
 	protected static Variable _scalar(Variable[] input1, Variable[] input2)  throws Exception {
-		float sum = 0F;
+		double sum = 0;
 		for (int i = 0; i < input1.length; i++) {
 			sum += input1[i].value * input2[i].value;
 		}
